+++ src/laminr/neuron_models/__init__.py
@@ -1,2 +1,2 @@
-from .monkey import monkey_v1
-from .simulated import simulated
+from .monkey_models import monkey_v1
+from .simulated_models import simulated
